Moving plot_vec into new binplotlib from individual plotting binaries
