+++ codebase/superdarn/src.bin/tk/plot/field_plot.1.12/makefile
@@ -6,12 +6,11 @@ include $(MAKECFG).$(SYSTEM)
 INCLUDE=-I$(IPATH)/base -I$(IPATH)/general -I$(IPATH)/analysis \
         -I$(IPATH)/imagery -I$(IPATH)/superdarn
 
-OBJS = plot_time.o plot_outline.o plot_vec.o plot_refvec.o make_fov.o \
+OBJS = plot_time.o plot_outline.o plot_refvec.o make_fov.o \
        expr.o field_plot.o
 
-SRC = hlpstr.h plot_time.c plot_time.h \
-      expr.c expr.h field_plot.c errstr.h \
-      plot_vec.c plot_vec.h make_fov.c make_fov.h \
+SRC = hlpstr.h plot_time.c plot_time.h expr.c expr.h \
+      field_plot.c errstr.h make_fov.c make_fov.h \
       plot_refvec.c plot_outline.c plot_outline.h \
       plot_refvec.h
