--- codebase/superdarn/src.bin/tk/plot/field_plot.1.12/plot_vec.h
@@ -1,40 +1,0 @@
-/* plot_vec.h
-   ========== 
-   Author: R.J.Barnes
-*/
-
-
-/*
- LICENSE AND DISCLAIMER
- 
- Copyright (c) 2012 The Johns Hopkins University/Applied Physics Laboratory
- 
- This file is part of the Radar Software Toolkit (RST).
- 
- RST is free software: you can redistribute it and/or modify
- it under the terms of the GNU Lesser General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- any later version.
- 
- RST is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU Lesser General Public License for more details.
- 
- You should have received a copy of the GNU Lesser General Public License
- along with RST.  If not, see <http://www.gnu.org/licenses/>.
- 
- 
- 
-*/
-
-
-
-
-
-void plot_vec(struct Plot *plot,struct RadarBeam *sbm,
-               struct GeoLocBeam *gbm,float latmin,int magflg,
-	      float xoff,float yoff,float wdt,float hgt,float sf,
-               int (*trnf)(int,void *,int,void *,void *data),void *data,
- 	       unsigned int(*cfn)(double,void *),void *cdata,
-               unsigned int gscol,unsigned char gsflg,float width,float rad);
