+++ codebase/superdarn/src.bin/tk/plot/grid_plot.1.12/grid_plot.c
@@ -1457,14 +1457,14 @@ int main(int argc,char *argv[]) {
 
     if (vecflg) {
       if (px==2) px+=10;
-      if (ortho) plot_vec(plot,px,1.8*apad,0,vmax,magflg,
-                        pad,pad,wdt-2*pad,hgt-2*pad,
-                        vsf,vradius,tfunc,marg,txtcol,0x0f,0.5,
-                        "Helvetica",10.0,fontdb,old_aacgm);
-      else plot_vec(plot,px,1.8*apad,0,vmax,magflg,
-                        pad,pad,wdt-2*pad,hgt-2*pad,
-                        vsf,vradius,MapStereographic,marg,txtcol,0x0f,0.5,
-                        "Helvetica",10.0,fontdb,old_aacgm);
+      if (ortho) plot_grid_vec(plot,px,1.8*apad,0,vmax,magflg,
+                               pad,pad,wdt-2*pad,hgt-2*pad,
+                               vsf,vradius,tfunc,marg,txtcol,0x0f,0.5,
+                               "Helvetica",10.0,fontdb,old_aacgm);
+      else plot_grid_vec(plot,px,1.8*apad,0,vmax,magflg,
+                         pad,pad,wdt-2*pad,hgt-2*pad,
+                         vsf,vradius,MapStereographic,marg,txtcol,0x0f,0.5,
+                         "Helvetica",10.0,fontdb,old_aacgm);
    }
 
   if (magflg || (!magflg && igrdflg)) {
